Refactors signature linking logic into a new hook and updates the bookmarks settings. The bookmarks settings tab now includes a new component with a dismissible warning.
- refactor(dialog): Extract signature linking logic into useLinkSignature hook
- feat(settings): Create BookmarksSettings component with dismissible warning
- chore(settings): Add migration for hideBookmarkWarning setting and bump version
- style(settings): Increase dialog height and improve scrollbars
