+++ assets/js/hooks/Mapper/mapRootProvider/version.ts
@@ -1,4 +1,4 @@
-export const STORED_SETTINGS_VERSION = 3;
+export const STORED_SETTINGS_VERSION = 4;
 
 export const LS_KEY_LEGASY = 'map-user-settings';
 export const LS_KEY = 'map-user-settings-v3';
